src/api/requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
azure-monitor-opentelemetry
opentelemetry-instrumentation-flask
dotenv